+++ src/Nazara/Core/ParameterList.cpp
@@ -91,7 +91,7 @@ bool NzParameterList::GetFloatParameter(const NzString& name, float* value) cons
 			return true;
 
 		case nzParameterType_Integer:
-			*value = it->second.value.intVal;
+			*value = static_cast<float>(it->second.value.intVal);
 			return true;
 
 		case nzParameterType_String:
@@ -99,7 +99,7 @@ bool NzParameterList::GetFloatParameter(const NzString& name, float* value) cons
 			double converted;
 			if (it->second.value.stringVal.ToDouble(&converted))
 			{
-				*value = converted;
+				*value = static_cast<float>(converted);
 				return true;
 			}
 
@@ -133,7 +133,7 @@ bool NzParameterList::GetIntegerParameter(const NzString& name, int* value) cons
 			return true;
 
 		case nzParameterType_Float:
-			*value = it->second.value.floatVal;
+			*value = static_cast<int>(it->second.value.floatVal);
 			return true;
 
 		case nzParameterType_Integer:
@@ -147,7 +147,7 @@ bool NzParameterList::GetIntegerParameter(const NzString& name, int* value) cons
 			{
 				if (converted <= std::numeric_limits<int>::max() && converted >= std::numeric_limits<int>::min())
 				{
-					*value = converted;
+					*value = static_cast<int>(converted);
 					return true;
 				}
 			}
